Changed raster draw method (slower but exact : fixed lots of bugs), and fixed bug of vector draw position linked to raster resolution

+++ src/gui/PanelViewer.hpp
@@ -205,7 +205,7 @@ protected:
 
 	void OnQuit(wxCommandEvent& event);
 
-	void Zoom(float zoomFactor, wxMouseEvent &event);
+	void Zoom(double zoomFactor, wxMouseEvent &event);
 
 	void UpdateIfTransformable();
